Refactor ForecastleApp to forecastle.go
