+++ pkg/apps/forecastle.go
@@ -1,0 +1,9 @@
+package apps
+
+// ForecastleApp struct that contains information about an app that is exposed to forecastle
+type ForecastleApp struct {
+	Name      string
+	Icon      string
+	Namespace string
+	URL       string
+}
